pyproject.toml
[tool.poetry]
name = "duplicated_video"
version = "0.1.0"
description = "Решение осеннего хакатона - 2024."
authors = [
    "Donskoi Andrei <donskoi.ae@gmail.com>",
    "Artem Medvedev <artiebears@mail.ru>",
    "Ali Ramazanov <ali_ramazanov_2000@mail.ru>",
    "Ivan Butakov <vanessbut@yandex.ru>",
    "Timofei Schcudro <t.shshudro@alumni.nsu.ru>",
]
packages = [
    { include = "duplicates/backend", from = "." },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.x"
numpy = "^1.26.4"
scikit-learn = "^1.3.2"
seaborn = "^0.13.0"
matplotlib = "^3.8.2"
tqdm = "^4.66.1"
torch = "^2.1.1"
plotly = "^5.18.0"
pytest = "^8.1.1"
fastapi = "0.115.0"
ruff = "^0.6.7"
uvicorn = "^0.30.6"
pytorchvideo = "^0.1.5"
faiss-cpu = "^1.8.0.post1"
jupyter = "^1.1.1"
torchvision = "^0.19.1"
pydantic = "2.9.2"
opencv-python = "^4.10.0.84"
tritonclient = {extras = ["all"], version = "^2.50.0"}
aio-pika = "9.4.3"
aiormq = "6.8.1"
annotated-types = "^0.7.0"
anyio = "4.6.0"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
click = "8.1.7"
greenlet = "3.1.1"
h11 = "0.14.0"
idna = "3.10"
msgpack = "1.1.0"
multidict = "6.1.0"
orjson = "3.10.7"
pamqp = "3.3.0"
prometheus-client = "0.21.0"
pydantic-settings = "2.5.2"
pydantic-core = "2.23.4"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
sniffio = "1.3.1"
sqlalchemy = "2.0.35"
starlette = "0.38.6"
starlette-context = "0.3.6"
typing-extensions = "4.12.2"
yarl = "1.13.0"
contextvars = "2.4"
python-multipart = "^0.0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80
lint.select = [
    "E",  # Ошибка стиля
    "W",  # Предупреждения по стилю
    "I",  # Корректировка импортов (аналог isort)
    "F",  # Аналог PyFlake
]